+++ dbaas/v1/client.go
@@ -53,15 +53,3 @@ func (c *DBaaSClient) Instances() InstanceService {
 func (c *DBaaSClient) Replicas() ReplicaService {
 	return &replicaService{client: c}
 }
-
-func (c *DBaaSClient) ParametersGroup() ParameterGroupService {
-	return &parameterGroupService{client: c}
-}
-
-func (c *DBaaSClient) Parameters() ParameterService {
-	return &parameterService{client: c}
-}
-
-func (c *DBaaSClient) Clusters() ClusterService {
-	return &clusterService{client: c}
-}
