+++ dbaas/v1/instance_types.go
@@ -25,7 +25,9 @@ type (
 		FamilyDescription string `json:"family_description"`
 		FamilySlug        string `json:"family_slug"`
 		Size              string `json:"size"`
-		CompatibleProduct string `json:"compatible_product"`
+		SKUSource         string `json:"sku_source"`
+		SKUReplica        string `json:"sku_replica"`
+		Status            string `json:"status,omitempty"`
 	}
 )
 
@@ -67,7 +69,7 @@ func (s *instanceTypeService) List(ctx context.Context, opts ListInstanceTypeOpt
 		s.client.newRequest,
 		s.client.GetConfig(),
 		http.MethodGet,
-		"/v2/instance-types",
+		"/v1/instance-types",
 		nil,
 		query,
 	)
@@ -84,7 +86,7 @@ func (s *instanceTypeService) Get(ctx context.Context, id string) (*InstanceType
 		s.client.newRequest,
 		s.client.GetConfig(),
 		http.MethodGet,
-		fmt.Sprintf("/v2/instance-types/%s", id),
+		fmt.Sprintf("/v1/instance-types/%s", id),
 		nil,
 		nil,
 	)
