+++ dbaas/v2/instances.go
@@ -7,18 +7,24 @@ import (
 	"net/url"
 	"strconv"
 	"strings"
+	"time"
 
 	mgc_http "github.com/MagaluCloud/mgc-sdk-go/internal/http"
 )
 
+const (
+	InstancePath   = "/v2/instances"
+	InstancePathID = InstancePath + "/%s"
+	SnapshotPath   = InstancePathID + "/snapshots"
+	SnapshotPathID = SnapshotPath + "/%s"
+)
+
 type (
-	InstanceStatus     string
-	InstanceGeneration string
-	VolumeType         string
-	AddressAccess      string
-	AddressType        string
-	SnapshotType       string
-	SnapshotStatus     string
+	InstanceStatus string
+	AddressAccess  string
+	AddressType    string
+	SnapshotType   string
+	SnapshotStatus string
 )
 
 const (
@@ -38,12 +44,6 @@ const (
 	InstanceStatusErrorDeleting    InstanceStatus = "ERROR_DELETING"
 	InstanceStatusMaintenance      InstanceStatus = "MAINTENANCE"
 	InstanceStatusMaintenanceError InstanceStatus = "MAINTENANCE_ERROR"
-)
-
-const (
-	VolumeTypeCloudNVME15K VolumeType = "CLOUD_NVME_15K"
-	VolumeTypeCloudNVME    VolumeType = "CLOUD_NVME"
-	VolumeTypeCloudHDD     VolumeType = "CLOUD_HDD"
 )
 
 const (
@@ -73,8 +73,8 @@ const (
 
 type (
 	Volume struct {
-		Size int        `json:"size"`
-		Type VolumeType `json:"type"`
+		Size int    `json:"size"`
+		Type string `json:"type"`
 	}
 
 	Address struct {
@@ -94,55 +94,36 @@ type (
 	}
 
 	InstanceVolumeRequest struct {
-		Size int        `json:"size"`
-		Type VolumeType `json:"type,omitempty"`
+		Size int    `json:"size"`
+		Type string `json:"type,omitempty"`
 	}
 
 	InstanceVolumeResizeRequest struct {
-		Size int        `json:"size"`
-		Type VolumeType `json:"type,omitempty"`
+		Size int    `json:"size"`
+		Type string `json:"type,omitempty"`
 	}
 
 	InstanceCreateRequest struct {
-		Name                string                       `json:"name"`
-		EngineID            *string                      `json:"engine_id,omitempty"`
-		InstanceTypeID      *string                      `json:"instance_type_id,omitempty"`
-		User                string                       `json:"user"`
-		Password            string                       `json:"password"`
-		Volume              InstanceVolumeRequest        `json:"volume"`
-		Parameters          *[]InstanceParametersRequest `json:"parameters,omitempty"`
-		BackupRetentionDays *int                         `json:"backup_retention_days,omitempty"`
-		BackupStartAt       *string                      `json:"backup_start_at,omitempty"`
+		Name                string                `json:"name"`
+		EngineID            *string               `json:"engine_id,omitempty"`
+		InstanceTypeID      *string               `json:"instance_type_id,omitempty"`
+		User                string                `json:"user"`
+		Password            string                `json:"password"`
+		Volume              InstanceVolumeRequest `json:"volume"`
+		ParameterGroupID    *string               `json:"parameter_group_id,omitempty"`
+		AvailabilityZone    *string               `json:"availability_zone,omitempty"`
+		BackupStartAt       *string               `json:"backup_start_at,omitempty"`
+		BackupRetentionDays *int                  `json:"backup_retention_days,omitempty"`
 	}
 
 	InstanceResizeRequest struct {
 		InstanceTypeID *string                      `json:"instance_type_id,omitempty"`
-		FlavorID       string                       `json:"flavor_id,omitempty"`
 		Volume         *InstanceVolumeResizeRequest `json:"volume,omitempty"`
 	}
 
 	DatabaseInstanceUpdateRequest struct {
-		Status              *InstanceStatusUpdate `json:"status,omitempty"`
-		BackupRetentionDays *int                  `json:"backup_retention_days,omitempty"`
-		BackupStartAt       *string               `json:"backup_start_at,omitempty"`
-	}
-
-	ReplicaDetailResponse struct {
-		ID                     string                       `json:"id"`
-		SourceID               string                       `json:"source_id"`
-		Name                   string                       `json:"name"`
-		EngineID               string                       `json:"engine_id"`
-		InstanceTypeID         string                       `json:"instance_type_id"`
-		Volume                 Volume                       `json:"volume"`
-		Addresses              []ReplicaAddressResponse     `json:"addresses"`
-		Status                 InstanceStatus               `json:"status"`
-		Generation             InstanceGeneration           `json:"generation"`
-		Parameters             []InstanceParametersResponse `json:"parameters"`
-		CreatedAt              string                       `json:"created_at"`
-		UpdatedAt              *string                      `json:"updated_at,omitempty"`
-		StartedAt              *string                      `json:"started_at,omitempty"`
-		FinishedAt             *string                      `json:"finished_at,omitempty"`
-		MaintenanceScheduledAt *string                      `json:"maintenance_scheduled_at,omitempty"`
+		BackupRetentionDays *int    `json:"backup_retention_days,omitempty"`
+		BackupStartAt       *string `json:"backup_start_at,omitempty"`
 	}
 
 	ReplicaAddressResponse struct {
@@ -157,17 +138,18 @@ type (
 	}
 
 	SnapshotDetailResponse struct {
-		ID            string                         `json:"id"`
-		Instance      SnapshotInstanceDetailResponse `json:"instance"`
-		Name          string                         `json:"name"`
-		Description   string                         `json:"description"`
-		Type          SnapshotType                   `json:"type"`
-		Status        SnapshotStatus                 `json:"status"`
-		AllocatedSize int                            `json:"allocated_size"`
-		CreatedAt     string                         `json:"created_at"`
-		StartedAt     *string                        `json:"started_at,omitempty"`
-		FinishedAt    *string                        `json:"finished_at,omitempty"`
-		UpdatedAt     *string                        `json:"updated_at,omitempty"`
+		ID               string                         `json:"id"`
+		Instance         SnapshotInstanceDetailResponse `json:"instance"`
+		Name             string                         `json:"name"`
+		Description      string                         `json:"description"`
+		Type             SnapshotType                   `json:"type"`
+		Status           SnapshotStatus                 `json:"status"`
+		AllocatedSize    int                            `json:"allocated_size"`
+		CreatedAt        time.Time                      `json:"created_at"`
+		StartedAt        *string                        `json:"started_at,omitempty"`
+		AvailabilityZone string                         `json:"availability_zone"`
+		FinishedAt       *string                        `json:"finished_at,omitempty"`
+		UpdatedAt        *time.Time                     `json:"updated_at,omitempty"`
 	}
 
 	SnapshotInstanceDetailResponse struct {
@@ -176,13 +158,13 @@ type (
 	}
 
 	SnapshotCreateRequest struct {
-		Name        string `json:"name"`
-		Description string `json:"description,omitempty"`
+		Name        string  `json:"name"`
+		Description *string `json:"description,omitempty"`
 	}
 
 	SnapshotUpdateRequest struct {
-		Name        string `json:"name,omitempty"`
-		Description string `json:"description,omitempty"`
+		Name        string  `json:"name,omitempty"`
+		Description *string `json:"description,omitempty"`
 	}
 
 	SnapshotResponse struct {
@@ -193,8 +175,8 @@ type (
 		Name                string                 `json:"name"`
 		InstanceTypeID      string                 `json:"instance_type_id"`
 		Volume              *InstanceVolumeRequest `json:"volume,omitempty"`
-		BackupRetentionDays int                    `json:"backup_retention_days,omitempty"`
-		BackupStartAt       string                 `json:"backup_start_at,omitempty"`
+		BackupRetentionDays *int                   `json:"backup_retention_days,omitempty"`
+		BackupStartAt       *string                `json:"backup_start_at,omitempty"`
 	}
 
 	ListSnapshotOptions struct {
@@ -203,21 +185,6 @@ type (
 		Type   *SnapshotType   `json:"type,omitempty"`
 		Status *SnapshotStatus `json:"status,omitempty"`
 	}
-)
-
-const (
-	InstanceGenerationG0B  InstanceGeneration = "G0B"
-	InstanceGenerationG1B  InstanceGeneration = "G1B"
-	InstanceGenerationG2B  InstanceGeneration = "G2B"
-	InstanceGenerationG3B  InstanceGeneration = "G3B"
-	InstanceGenerationG4B  InstanceGeneration = "G4B"
-	InstanceGenerationG5B  InstanceGeneration = "G5B"
-	InstanceGenerationG6B  InstanceGeneration = "G6B"
-	InstanceGenerationG7B  InstanceGeneration = "G7B"
-	InstanceGenerationG8B  InstanceGeneration = "G8B"
-	InstanceGenerationG9B  InstanceGeneration = "G9B"
-	InstanceGenerationG10B InstanceGeneration = "G10B"
-	InstanceGenerationG1   InstanceGeneration = "G1"
 )
 
 type InstanceStatusUpdate string
@@ -273,7 +240,7 @@ type (
 		DeleteSnapshot(ctx context.Context, instanceID, snapshotID string) error
 
 		// RestoreSnapshot creates a new instance from a snapshot.
-		RestoreSnapshot(ctx context.Context, snapshotID string, req RestoreSnapshotRequest) (*InstanceResponse, error)
+		RestoreSnapshot(ctx context.Context, instanceID, snapshotID string, req RestoreSnapshotRequest) (*InstanceResponse, error)
 	}
 
 	instanceService struct {
@@ -307,25 +274,27 @@ type (
 	}
 
 	InstanceDetail struct {
-		ID                     string                       `json:"id"`
-		Name                   string                       `json:"name"`
-		EngineID               string                       `json:"engine_id"`
-		DatastoreID            string                       `json:"datastore_id"`
-		FlavorID               string                       `json:"flavor_id"`
-		InstanceTypeID         string                       `json:"instance_type_id"`
-		Volume                 Volume                       `json:"volume"`
-		Addresses              []Address                    `json:"addresses"`
-		Status                 InstanceStatus               `json:"status"`
-		Generation             InstanceGeneration           `json:"generation"`
-		Parameters             []InstanceParametersResponse `json:"parameters"`
-		BackupRetentionDays    int                          `json:"backup_retention_days"`
-		BackupStartAt          string                       `json:"backup_start_at"`
-		CreatedAt              string                       `json:"created_at"`
-		UpdatedAt              *string                      `json:"updated_at,omitempty"`
-		StartedAt              *string                      `json:"started_at,omitempty"`
-		FinishedAt             *string                      `json:"finished_at,omitempty"`
-		MaintenanceScheduledAt *string                      `json:"maintenance_scheduled_at,omitempty"`
-		Replicas               []ReplicaDetailResponse      `json:"replicas,omitempty"`
+		ID                     string                  `json:"id"`
+		Name                   string                  `json:"name"`
+		EngineID               string                  `json:"engine_id"`
+		DatastoreID            string                  `json:"datastore_id"`
+		FlavorID               string                  `json:"flavor_id"`
+		InstanceTypeID         string                  `json:"instance_type_id"`
+		Volume                 Volume                  `json:"volume"`
+		Addresses              []Address               `json:"addresses"`
+		Status                 InstanceStatus          `json:"status"`
+		Generation             string                  `json:"generation"`
+		ApplyParametersPending bool                    `json:"apply_parameters_pending"`
+		ParameterGroupID       string                  `json:"parameter_group_id"`
+		AvailabilityZone       string                  `json:"availability_zone"`
+		BackupRetentionDays    int                     `json:"backup_retention_days"`
+		BackupStartAt          string                  `json:"backup_start_at"`
+		CreatedAt              string                  `json:"created_at"`
+		UpdatedAt              *string                 `json:"updated_at,omitempty"`
+		StartedAt              *string                 `json:"started_at,omitempty"`
+		FinishedAt             *string                 `json:"finished_at,omitempty"`
+		MaintenanceScheduledAt *string                 `json:"maintenance_scheduled_at,omitempty"`
+		Replicas               []ReplicaDetailResponse `json:"replicas,omitempty"`
 	}
 )
 
@@ -370,7 +339,7 @@ func (s *instanceService) List(ctx context.Context, opts ListInstanceOptions) ([
 		s.client.newRequest,
 		s.client.GetConfig(),
 		http.MethodGet,
-		"/v1/instances",
+		InstancePath,
 		nil,
 		query,
 	)
@@ -394,7 +363,7 @@ func (s *instanceService) Get(ctx context.Context, id string, opts GetInstanceOp
 		s.client.newRequest,
 		s.client.GetConfig(),
 		http.MethodGet,
-		fmt.Sprintf("/v1/instances/%s", id),
+		fmt.Sprintf(InstancePathID, id),
 		nil,
 		query,
 	)
@@ -408,7 +377,7 @@ func (s *instanceService) Create(ctx context.Context, req InstanceCreateRequest)
 		s.client.newRequest,
 		s.client.GetConfig(),
 		http.MethodPost,
-		"/v1/instances",
+		InstancePath,
 		req,
 		nil,
 	)
@@ -422,7 +391,7 @@ func (s *instanceService) Delete(ctx context.Context, id string) error {
 		s.client.newRequest,
 		s.client.GetConfig(),
 		http.MethodDelete,
-		fmt.Sprintf("/v1/instances/%s", id),
+		fmt.Sprintf(InstancePathID, id),
 		nil,
 		nil,
 	)
@@ -436,7 +405,7 @@ func (s *instanceService) Update(ctx context.Context, id string, req DatabaseIns
 		s.client.newRequest,
 		s.client.GetConfig(),
 		http.MethodPatch,
-		fmt.Sprintf("/v1/instances/%s", id),
+		fmt.Sprintf(InstancePathID, id),
 		req,
 		nil,
 	)
@@ -450,7 +419,7 @@ func (s *instanceService) Resize(ctx context.Context, id string, req InstanceRes
 		s.client.newRequest,
 		s.client.GetConfig(),
 		http.MethodPost,
-		fmt.Sprintf("/v1/instances/%s/resize", id),
+		fmt.Sprintf(InstancePathID+"/resize", id),
 		req,
 		nil,
 	)
@@ -464,7 +433,7 @@ func (s *instanceService) Start(ctx context.Context, id string) (*InstanceDetail
 		s.client.newRequest,
 		s.client.GetConfig(),
 		http.MethodPost,
-		fmt.Sprintf("/v1/instances/%s/start", id),
+		fmt.Sprintf(InstancePathID+"/start", id),
 		nil,
 		nil,
 	)
@@ -478,7 +447,7 @@ func (s *instanceService) Stop(ctx context.Context, id string) (*InstanceDetail,
 		s.client.newRequest,
 		s.client.GetConfig(),
 		http.MethodPost,
-		fmt.Sprintf("/v1/instances/%s/stop", id),
+		fmt.Sprintf(InstancePathID+"/stop", id),
 		nil,
 		nil,
 	)
@@ -506,7 +475,7 @@ func (s *instanceService) ListSnapshots(ctx context.Context, instanceID string, 
 		s.client.newRequest,
 		s.client.GetConfig(),
 		http.MethodGet,
-		fmt.Sprintf("/v1/instances/%s/snapshots", instanceID),
+		fmt.Sprintf(SnapshotPath, instanceID),
 		nil,
 		query,
 	)
@@ -524,7 +493,7 @@ func (s *instanceService) CreateSnapshot(ctx context.Context, instanceID string,
 		s.client.newRequest,
 		s.client.GetConfig(),
 		http.MethodPost,
-		fmt.Sprintf("/v1/instances/%s/snapshots", instanceID),
+		fmt.Sprintf(SnapshotPath, instanceID),
 		req,
 		nil,
 	)
@@ -537,7 +506,7 @@ func (s *instanceService) GetSnapshot(ctx context.Context, instanceID, snapshotI
 		s.client.newRequest,
 		s.client.GetConfig(),
 		http.MethodGet,
-		fmt.Sprintf("/v1/instances/%s/snapshots/%s", instanceID, snapshotID),
+		fmt.Sprintf(SnapshotPathID, instanceID, snapshotID),
 		nil,
 		nil,
 	)
@@ -550,7 +519,7 @@ func (s *instanceService) UpdateSnapshot(ctx context.Context, instanceID, snapsh
 		s.client.newRequest,
 		s.client.GetConfig(),
 		http.MethodPatch,
-		fmt.Sprintf("/v1/instances/%s/snapshots/%s", instanceID, snapshotID),
+		fmt.Sprintf(SnapshotPathID, instanceID, snapshotID),
 		req,
 		nil,
 	)
@@ -563,20 +532,20 @@ func (s *instanceService) DeleteSnapshot(ctx context.Context, instanceID, snapsh
 		s.client.newRequest,
 		s.client.GetConfig(),
 		http.MethodDelete,
-		fmt.Sprintf("/v1/instances/%s/snapshots/%s", instanceID, snapshotID),
+		fmt.Sprintf(SnapshotPathID, instanceID, snapshotID),
 		nil,
 		nil,
 	)
 }
 
 // RestoreSnapshot creates a new instance from a snapshot.
-func (s *instanceService) RestoreSnapshot(ctx context.Context, snapshotID string, req RestoreSnapshotRequest) (*InstanceResponse, error) {
+func (s *instanceService) RestoreSnapshot(ctx context.Context, instanceID, snapshotID string, req RestoreSnapshotRequest) (*InstanceResponse, error) {
 	return mgc_http.ExecuteSimpleRequestWithRespBody[InstanceResponse](
 		ctx,
 		s.client.newRequest,
 		s.client.GetConfig(),
 		http.MethodPost,
-		fmt.Sprintf("/v1/snapshots/%s/restores", snapshotID),
+		fmt.Sprintf(SnapshotPathID+"/restore", instanceID, snapshotID),
 		req,
 		nil,
 	)
